codebase refactor >> separate model for user and admin
