+++ src/pages/user/Faqs.jsx
